fix shaking XYWnd::XY_DrawAxis position

+++ radiant/xywindow.cpp
@@ -1403,8 +1403,8 @@ void XYWnd::XY_DrawAxis( void ){
 		const char g_AxisName[3] = { 'X', 'Y', 'Z' };
 		const int nDim1 = ( m_viewType == YZ ) ? 1 : 0;
 		const int nDim2 = ( m_viewType == XY ) ? 1 : 2;
-		const int w = ( m_nWidth / 2 / m_fScale );
-		const int h = ( m_nHeight / 2 / m_fScale );
+		const float w = ( m_nWidth / 2 / m_fScale );
+		const float h = ( m_nHeight / 2 / m_fScale );
 
 		Vector3 colourX = ( m_viewType == YZ ) ? g_xywindow_globals.AxisColorY : g_xywindow_globals.AxisColorX;
 		Vector3 colourY = ( m_viewType == XY ) ? g_xywindow_globals.AxisColorY : g_xywindow_globals.AxisColorZ;
